Remove commands.php. Move last 2 funcs to lib_inventory
